requirements.txt
asgiref==3.8.1
Django==5.2
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
whitenoise==6.9.0
